refactor code into separate file to retain readability
