+++ Server/src/com/benberi/cadesim/server/service/GameServerBootstrap.java
@@ -12,12 +12,7 @@ import org.apache.commons.cli.HelpFormatter;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.time.ZonedDateTime;
-import java.util.ArrayList;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ThreadLocalRandom;
@@ -94,60 +89,8 @@ public class GameServerBootstrap {
     public static void main(String[] args) throws NumberFormatException, InterruptedException{
         ServerContext.log("Welcome to " + Constants.name + " (version " + Constants.VERSION + ")" + ".");
 
-        // add a few sec delay before doing anything to give any
-        // previous instances a chance to exit
-        Thread.sleep(2000);
-
-        // check if the lockdir is "ours"
-        boolean lastUpdateWasOurs = false;
-        String idfilename =
-                Constants.AUTO_UPDATING_LOCK_DIRECTORY_NAME +
-                System.getProperty("file.separator") +
-                Constants.AUTO_UPDATING_ID_FILE_NAME;
-        String idfilecontent;
-        try {
-            idfilecontent = new String(
-                Files.readAllBytes(Paths.get(idfilename))
-             );
-            if (idfilecontent.equals(String.join(" ", ServerConfiguration.getArgs()))) {
-                lastUpdateWasOurs = true;
-            }
-        } catch (IOException e) {
-            // pass, couldn't open idfile so assume wasn't ours
-        }
-
-        // if last update was ours, clean up
-        if (lastUpdateWasOurs) {
-            ArrayList<String> toDelete = new ArrayList<>();
-            toDelete.add("getdown.txt");
-            toDelete.add("version.txt");
-            toDelete.add("digest.txt");
-            toDelete.add("digest2.txt");
-            toDelete.add(
-                Constants.AUTO_UPDATING_LOCK_DIRECTORY_NAME +
-                System.getProperty("file.separator") +
-                Constants.AUTO_UPDATING_ID_FILE_NAME
-            );
-            toDelete.add(
-                Constants.AUTO_UPDATING_LOCK_DIRECTORY_NAME
-            );
-            boolean deleteSucceeded = true;
-            for (String s : toDelete) {
-                File f = new File(s);
-                if (f.delete()) {
-                    ServerContext.log("deleted " + f.getPath() + " on startup.");
-                }
-                else
-                {
-                    ServerContext.log("couldn't delete " + f.getPath() + " on startup");
-                    deleteSucceeded = false;
-                }
-            }
-
-            if (!deleteSucceeded) {
-                ServerContext.log("Error: delete of at least one update file failed.");
-            }
-        }
+        // clean up after previous update session if needed
+        Updater.cleanupAfterRestart();
 
         // set up maps
         ServerConfiguration.loadAvailableMaps();
